Check for missing original dll

+++ EdiabasLib/EdiabasLibConfigTool/Patch.cs
@@ -340,29 +340,39 @@ namespace EdiabasLibConfigTool
 #endif
                 // 32 bit
                 string dllFile32 = Path.Combine(dirName, Api32DllName);
-                string dllFile32Backup = Path.Combine(dirName, Api32DllBackupName);
-                if (!File.Exists(dllFile32Backup))
+                if (File.Exists(dllFile32))
                 {
-                    if (IsOriginalDll(dllFile32))
+                    string dllFile32Backup = Path.Combine(dirName, Api32DllBackupName);
+                    if (!File.Exists(dllFile32Backup))
+                    {
+                        if (IsOriginalDll(dllFile32))
+                        {
+                            sr.Append("\r\n");
+                            sr.Append(string.Format(Resources.Strings.PatchCreateBackupFile, Api32DllBackupName));
+                            File.Copy(dllFile32, dllFile32Backup, false);
+                        }
+                    }
+                    else
                     {
                         sr.Append("\r\n");
-                        sr.Append(string.Format(Resources.Strings.PatchCreateBackupFile, Api32DllBackupName));
-                        File.Copy(dllFile32, dllFile32Backup, false);
+                        sr.Append(string.Format(Resources.Strings.PatchBackupFileExisting, Api32DllBackupName));
                     }
+
+                    if (!IsOriginalDll(dllFile32Backup))
+                    {
+                        sr.Append("\r\n");
+                        sr.Append(string.Format(Resources.Strings.PatchNoValidBackupFile, Api32DllName));
+                        return false;
+                    }
+
+                    File.Copy(sourceDll32, dllFile32, true);
                 }
                 else
                 {
                     sr.Append("\r\n");
-                    sr.Append(string.Format(Resources.Strings.PatchBackupFileExisting, Api32DllBackupName));
-                }
-
-                if (!IsOriginalDll(dllFile32Backup))
-                {
-                    sr.Append("\r\n");
-                    sr.Append(string.Format(Resources.Strings.PatchNoValidBackupFile, Api32DllName));
+                    sr.Append(string.Format(Resources.Strings.PatchOriginalApiDllMissing, Api32DllName));
                     return false;
                 }
-                File.Copy(sourceDll32, dllFile32, true);
 
                 // 64 bit
                 string dllFile64 = Path.Combine(dirName, Api64DllName);
@@ -390,7 +400,13 @@ namespace EdiabasLibConfigTool
                         sr.Append(string.Format(Resources.Strings.PatchNoValidBackupFile, Api64DllName));
                         return false;
                     }
+
                     File.Copy(sourceDll64, dllFile64, true);
+                }
+                else
+                {
+                    sr.Append("\r\n");
+                    sr.Append(string.Format(Resources.Strings.PatchOriginalApiDllMissing, Api64DllName));
                 }
 
                 string sourceConfig = Path.Combine(sourceDir, ConfigFileName);
